requirements.txt
apprise==1.6.0
PyYaml==6.0.1